--- src/city/components/EditModal.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-
-export const EditModal = () => {
-    //TODO
-    return(<>Edit modal</>)
-}
