--- src/country/components/EditModal.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-
-export const EditModal = () => {
-    //TODO
-    return(<>Edit modal</>)
-}
