+++ src/hotel/types.ts
@@ -1,11 +1,6 @@
 export type Hotel = {
+    id: number,
     name?: string,
     address?: string,
     price?: number
-}
-
-export type HotelModalProps = {
-    hotel: Hotel,
-    open: boolean,
-    handleClose: any
 }
